Create a series of tests for self-testing.
The tester has some tests of its own functionality.
We create an alternate set of tests starting with test2_
to ensure that the test runner has something to run
when doing its test.

This also ensure that the testing statistics won't change
and we can test them.

Each of these tests has a test2_ test -- and a test_ test
that gets the result from the tester.

We capture the standard output to get the text printed
using print() -- then we compare that to the expected result.