Added Setup and Teardown functions.
The test runner now has setup and teardown
functions. This is useful for our tests as many of the
test had to run the test runner and get the output.

Now we just do that once, storing the result in a
global variable. This simplifies many of the tests.

Added some tests of the setup and teardown itself.
These are difficult since they can lead to an infinite
loop where the setup runs the test runner -- which runs setup.

Teardown has its own challenge -- since it's not supposed to run
until all tests are done, so how do you have a test check to see
if it has run?

My answer is to assert that it ran at the end.

Add setup and teardown.

Add setup and teardown tests.

The tests work, but they have to run the whole test engine to work.