Update start.py
Here's my attempt at implementing multiple parenthood. Parents now needs to be defined as a list. But that means you can have as many as you like, and when we go searching for methods we search in each one. It seems that when we search for methods we're going alphabetically and grabbing the first one we find. No time to account for that presently, but this was an interesting exercise that was healthily befuddling.